+++ src/docs/HISTORY.md
@@ -9,6 +9,8 @@ v1.3.1
 - Made versioning regex get escaped, like Python's re.escape() function.
 - Added support for version switching of clips in .otio timeline.   You need
   to be stopped at a certain clip to change it.
+- Added 'm' hotkey to mark the in and out points of a clip in the .otio
+  timeline.
 - Upgraded to OpenEXR v3.3.2.
 - Added OpenEXR's headers Compression, Compression Num. Scanlines,
   and Is Deep, Is Lossy.